Move proxies to separate file.
