+++ tests/suites/proxies.js
@@ -1,0 +1,139 @@
+/*!
+ * Copyright 2024 Digital Bazaar, Inc.
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+import crypto from 'node:crypto';
+
+export function createProxy({original, stubs}) {
+  if(typeof original === 'object') {
+    throw new Error(`Expected parameter original to be an object received ` +
+    `${typeof original}`);
+  }
+  return new Proxy(original, {
+    get(target, prop) {
+      if(stubs[prop]) {
+        return stubs[prop];
+      }
+      return Reflect.get(...arguments);
+    }
+  });
+}
+
+/**
+ * The major jsonld api suites use is canonize.
+ * This function intercepts calls on canonize and
+ * pass safe: false allowing for invalid jsonld to
+ * be issued.
+ *
+ * @param {object} suite - A DataIntegrityProof.
+ *
+ * @returns {Proxy<object>} Returns a proxy of the proof.
+ */
+export function unsafeProxy(suite) {
+  if(typeof suite !== 'object') {
+    return suite;
+  }
+  // if the suite has a cryptosuite object proxy it
+  if(suite._cryptosuite) {
+    suite._cryptosuite = new Proxy(suite._cryptosuite, {
+      get(target, prop) {
+        if(prop === 'canonize') {
+          return function(doc, options) {
+            return target.canonize(doc, {...options, safe: false});
+          };
+        }
+        return Reflect.get(...arguments);
+      }
+    });
+  }
+  return new Proxy(suite, {
+    get(target, prop) {
+      if(prop === 'canonize') {
+        return function(doc, options) {
+          return target.canonize(doc, {...options, safe: false});
+        };
+      }
+      return Reflect.get(...arguments);
+    }
+  });
+}
+
+export function invalidHashProxy({
+  suiteName,
+  keyType,
+  suite,
+}) {
+  if(typeof suite !== 'object') {
+    return suite;
+  }
+  if(suite._cryptosuite) {
+    if(suiteName !== 'ecdsa-rdfc-2019') {
+      throw new Error(`Unsupported suite ${suiteName}`);
+    }
+    suite._cryptosuite = new Proxy(suite._cryptosuite, {
+      get(target, prop) {
+        if(prop === 'createVerifyData') {
+          return async function({
+            cryptosuite, document, proof,
+            documentLoader, dataIntegrityProof
+          } = {}) {
+            // this switch the hash to the wrong hash for that keyType
+            const algorithm = (keyType === 'P-256') ? 'sha384' : 'sha256';
+            const c14nOptions = {
+              documentLoader,
+              safe: true,
+              base: null,
+              skipExpansion: false,
+              messageDigestAlgorithm: algorithm
+            };
+
+            // await both c14n proof hash and c14n document hash
+            const [proofHash, docHash] = await Promise.all([
+              // canonize and hash proof
+              _canonizeProof(proof, {
+                document, cryptosuite, dataIntegrityProof, c14nOptions
+              }).then(c14nProofOptions => sha({
+                algorithm,
+                string: c14nProofOptions
+              })),
+              // canonize and hash document
+              cryptosuite.canonize(document, c14nOptions).then(
+                c14nDocument => sha({algorithm, string: c14nDocument}))
+            ]);
+            // concatenate hash of c14n proof options and hash of c14n document
+            return _concat(proofHash, docHash);
+          };
+        }
+        return Reflect.get(...arguments);
+      }
+    });
+  }
+  return suite;
+}
+
+function _concat(b1, b2) {
+  const rval = new Uint8Array(b1.length + b2.length);
+  rval.set(b1, 0);
+  rval.set(b2, b1.length);
+  return rval;
+}
+
+export async function sha({algorithm, string}) {
+  return new Uint8Array(crypto.createHash(algorithm).update(string).digest());
+}
+
+async function _canonizeProof(proof, {
+  document, cryptosuite, dataIntegrityProof, c14nOptions
+}) {
+  // `proofValue` must not be included in the proof options
+  proof = {
+    '@context': document['@context'],
+    ...proof
+  };
+  dataIntegrityProof.ensureSuiteContext({
+    document: proof, addSuiteContext: true
+  });
+  delete proof.proofValue;
+  return cryptosuite.canonize(proof, c14nOptions);
+}
+
